// src/app/models/task/properties/map/productCategory.ts
export enum ProductCategory { 
  
  FOOD = 'FOOD',
  PRODUCTS = 'PRODUCTS',
  FLOWERS = 'FLOWERS',
  HOUSEHOLD_GOODS = 'HOUSEHOLD_GOODS',
  APPLIANCES = 'APPLIANCES',
  STATIONERY_AND_BOOKS = 'STATIONERY_AND_BOOKS',
  BUILDING_MATERIALS = 'BUILDING_MATERIALS',
  CONSTRUCTION_AND_GARDEN_TOOLS = 'CONSTRUCTION_AND_GARDEN_TOOLS',
  COUNTRY_TOOLS = 'COUNTRY_TOOLS',
  MEDICAL_SUPPLIES = 'MEDICAL_SUPPLIES',
  ALCOHOL_AND_TOBACCO = 'ALCOHOL_AND_TOBACCO',
  GOODS_FOR_PETS = 'GOODS_FOR_PETS',
  COSMETICS_AND_EVERYTHING_FOR_BEAUTY = 'COSMETICS_AND_EVERYTHING_FOR_BEAUTY',
  EVERYTHING_FOR_CARS = 'EVERYTHING_FOR_CARS',
  EVERYTHING_FOR_OUTDOOR_ACTIVITIES = 'EVERYTHING_FOR_OUTDOOR_ACTIVITIES',
  CHILDRENS_PRODUCTS = 'CHILDRENS_PRODUCTS',
  OTHER = 'OTHER'

/*
  Еда (Из заведений)
  Продукты (Из магазина)
  Цветы
  Бытовые товары
  Бытовая техника
  Канцелярия и книги
  Стройматериалы
  Строительные и садовые инструменты
  Для дачи
  Медицинские товары
  Алкоголь и табак
  Товары для животных
  Косметика и всё для красоты
  Всё для автомобилей
  Всё для активного отдыха
  Товары для детей
  Прочее
*/
}