// src/app/models/task/properties/map/distanceType.ts
export enum DistanceType {

    FROM_0_TO_5 = 'FROM_0_TO_5',
    FROM_5_TO_10 = 'FROM_5_TO_10',
    FROM_10_TO_20 = 'FROM_10_TO_20',
    FROM_20_TO_30 = 'FROM_20_TO_30',
    FROM_30_TO_40 = 'FROM_30_TO_40',
    FROM_40_TO_50 = 'FROM_40_TO_50',
    FROM_50_TO_60 = 'FROM_50_TO_60',
    FROM_60 = 'FROM_60'

}