// src/app/models/task/properties/map/transportCategory.ts
export enum TransportCategory {
  ON_FOOT = 'ON_FOOT',
  BIKE_MOTORCYCLE_MOPED = 'BIKE_MOTORCYCLE_MOPED',
  CAR = 'CAR',
  PASSENGER_VAN = 'PASSENGER_VAN',
  TRUCK_TO_2 = 'TRUCK_TO_2',
  TRUCK_FROM_2_TO_5 = 'TRUCK_FROM_2_TO_5',
  TRUCK_FROM_5_TO_10 = 'TRUCK_FROM_5_TO_10',
  TRUCK_FROM_10 = 'TRUCK_FROM_10',

  MINIBUS = 'MINIBUS',
  BUS = 'BUS',
  ANOTHER_TYPE_TRANSPORT = 'ANOTHER_TYPE_TRANSPORT'

}