// src/app/models/task/properties/map/weightCategory.ts
export enum WeightCategory {
  FROM_0001_TO_4990_GR = 'FROM_0001_TO_4990_GR',
  FROM_5000_TO_9990_GR = 'FROM_5000_TO_9990_GR',
  FROM_10_000_TO_19_990_GR = 'FROM_10_000_TO_19_990_GR',
  FROM_20_000_TO_49_990_GR = 'FROM_20_000_TO_49_990_GR',
  FROM_50_000_TO_99_990_GR = 'FROM_50_000_TO_99_990_GR',
  FROM_100_000_TO_499_990_GR = 'FROM_100_000_TO_499_990_GR',
  FROM_500_000_TO_999_990_GR = 'FROM_500_000_TO_999_990_GR',

  FROM_1_000_TO_1_990_KG = 'FROM_1_000_TO_1_990_KG',
  FROM_2_000_TO_4_990_KG = 'FROM_2_000_TO_4_990_KG',
  FROM_5_000_TO_9_990_KG = 'FROM_5_000_TO_9_990_KG',
  FROM_10_000_KG = 'FROM_10_000_KG',

}