// src/app/models/chat/arbitration/chatArbitrationStatus.ts
export enum ChatArbitrationStatus {
  IN_PROGRESS = 'IN_PROGRESS',
  ADMIN_REQUEST = 'ADMIN_REQUEST',
  ADMIN_JOINED = 'ADMIN_JOINED',
  SUCCESS = 'SUCCESS',
  SUCCESS_IN_FAVOR_CUSTOMER = 'SUCCESS_IN_FAVOR_CUSTOMER',
  SUCCESS_IN_FAVOR_PERFORMER = 'SUCCESS_IN_FAVOR_PERFORMER'
}

export const FINISHED_CHAT_ARBITRATION_STATUSES = [
  ChatArbitrationStatus.SUCCESS,
  ChatArbitrationStatus.SUCCESS_IN_FAVOR_CUSTOMER, 
  ChatArbitrationStatus.SUCCESS_IN_FAVOR_PERFORMER
];